⚡️ Even more packing

+++ src/Vault.sol
@@ -202,13 +202,13 @@ contract Vault is ERC20, Auth {
 
     /// @notice A timestamp representing when the first harvest in the most recent harvest window occurred.
     /// @dev May be equal to lastHarvest if there was/has only been one harvest in the most last/current window.
-    uint128 public lastHarvestWindowStart;
+    uint64 public lastHarvestWindowStart;
 
     /// @notice A timestamp representing when the most recent harvest occurred.
-    uint128 public lastHarvest;
+    uint64 public lastHarvest;
 
     /// @notice The amount of locked profit at the end of the last harvest.
-    uint256 public maxLockedProfit;
+    uint128 public maxLockedProfit;
 
     /*///////////////////////////////////////////////////////////////
                       WITHDRAWAL QUEUE STORAGE
@@ -401,7 +401,7 @@ contract Vault is ERC20, Auth {
         // If this is the first harvest after the last window:
         if (block.timestamp >= lastHarvest + harvestDelay) {
             // Set the harvest window's start timestamp.
-            lastHarvestWindowStart = uint128(block.timestamp);
+            lastHarvestWindowStart = uint64(block.timestamp);
         } else {
             // We know this harvest is not the first in the window so we need to ensure it's within it.
             require(block.timestamp <= lastHarvestWindowStart + harvestWindow, "BAD_HARVEST_TIME");
@@ -432,10 +432,10 @@ contract Vault is ERC20, Auth {
         balanceOfStrategy[strategy] = balanceThisHarvest;
 
         // Update the max amount of locked profit.
-        maxLockedProfit = profitAccrued - feesAccrued;
+        maxLockedProfit = uint128(profitAccrued - feesAccrued);
 
         // Update the last harvest timestamp.
-        lastHarvest = uint128(block.timestamp);
+        lastHarvest = uint64(block.timestamp);
 
         // Get the next harvest delay.
         uint128 newHarvestDelay = nextHarvestDelay;
